feat: move modal to shared
